product price from int to float

+++ coffee-order-terminal/app/src/main/java/com/feup/coffee_order_terminal/models/Product.kt
@@ -3,7 +3,7 @@ package com.feup.coffee_order_terminal.models
 data class Product(
     val _id: String,
     val name: String,
-    val price: Int,
+    val price: Float,
     val imgURL: String,
     val category: Category
 )
